Fix a data race cause an assertion failure.
There is insufficient atomicity within Dhat's methods. So it's possible
to have a sequence like this:
- `System.dealloc` frees `p`
- `System.alloc` allocates (recycles) `p`
- dhat-rs accounts for the `alloc`
- dhat-rs accounts for the `dealloc`

But dhat-rs gets confused and asserts on the third step.

The solution is to move the `System.{alloc,realloc,dealloc}` calls
within the scope of the `TRI_GLOBALS` mutex (so long as we're not
ignoring allocs). That avoids the bad ordering mentioned above.

Fixes #4.